Agents/requirements.txt
crewai
crewai-tools
load_dotenv
docx2txt
fastapi
uvicorn
requests